Changes int types to xint_fastx_t

+++ src/semi_positive_filter.cc
@@ -9,7 +9,7 @@ SemiPositiveFilter::operator()(arma::Mat<int> const& a) {
 	constexpr double tol = 1e-10;
 	_double_mat = arma::conv_to<arma::mat>::from(a);
 	bool result = arma::eig_sym(_eigens, _double_mat);
-	for(uint32_t i = 0; result && i < _eigens.n_elem; ++i) {
+	for(uint_fast16_t i = 0; result && i < _eigens.n_elem; ++i) {
 		result = _eigens(i) > -tol;
 	}
 	return result;
